Adds more inline completion tests
